server/requirements.txt
wheel
numpy
Django
djangorestframework
tzdata
context-compose
context.api
django-leaflet
influxdb-client
requests